move dummy out of mod
